CIV-11185 Add common changes
