--- src/main/common/form/models/fee.ts
@@ -1,4 +1,0 @@
-export enum FeeType {
-  HEARING = 'HEARING',
-  CLAIM_ISSUED = 'CLAIMISSUED',
-}
